Split device module into two parts
One part stays in the physical component, one part goes up a level.
