+++ src/engine/strat_engine/tests/real.rs
@@ -15,7 +15,7 @@ use devicemapper::{Bytes, IEC, Sectors};
 use super::logger::init_logger;
 use super::util::clean_up;
 
-use super::super::physical::wipe_sectors;
+use super::super::device::wipe_sectors;
 
 pub struct RealTestDev {
     path: PathBuf,
